+++ sharedPlugins/serverManager/index.js
@@ -10,7 +10,6 @@ module.exports = class remoteCommands {
 		this.config = mergedConfig;
 		this.socket = extras.socket;
 		const pluginManager = require("lib/manager/pluginManager.js");
-		const configManager = require("lib/manager/configManager.js");
 		// const modManager = require("lib/manager/modManager.js")(this.config);
 		this.socket.on("hello", () => {
 			this.socket.emit("registerServerManager");
@@ -22,13 +21,6 @@ module.exports = class remoteCommands {
 			// TODO: Implement this function separately from listMods (which console.logs it)
 			// let mods = modManager.getMods()
 			// callback(mods);
-		});
-		this.socket.on("serverManagerGetConfig", async (data, callback) => {
-			let config = await configManager.getConfig(data.instance);
-			callback(config);
-		});
-		this.socket.on("serverManagerEditConfig", async (data, callback) => {
-			callback(configManager.editConfig(data.instance, data.entry, data.value));
 		});
 		this.socket.on("serverManagerGetPlugins", async (data, callback) => {
 			callback(await pluginManager.getPlugins());