+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Version 2.0.0
   Breaks discordChat, playerManager, serverSelect, trainTeleports and
   clusterioModel.
 - Instance names can no longer be invalid Windows file names.
+- Removed config management from the command line and the server manager.
 
 
 Version 1.2.1